requirements.txt
Flask==1.1.2
flask-sqlalchemy==2.4.4
sqlalchemy-serializer==1.3.4.2
flask-restplus==0.13.0
flask-cors==3.0.8
mysqlclient==2.0.1
werkzeug==0.16.0
bcrypt==3.1.7
python-dateutil==2.8.1
PyJWT